changed random field to act as a container.  changed learning to use floats instead of doubles
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@18305 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 8957f2e9df5c4428ed3793f245e7e117b1c3015e

+++ sandbox/functional_m3n/include/functional_m3n/m3n_model.h
@@ -180,9 +180,9 @@ class M3NModel
      * \return The functional gradient residual
      */
     // --------------------------------------------------------------
-    double calcFuncGradResidual(const double truncation_param,
-                                const unsigned int clique_order,
-                                const unsigned int nbr_mode_label);
+    float calcFuncGradResidual(const double truncation_param,
+                               const unsigned int clique_order,
+                               const unsigned int nbr_mode_label);
 
     // --------------------------------------------------------------
     /*!
